server/requirements.txt
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
click==8.2.0
cssselect2==0.8.0
defusedxml==0.7.1
Flask==3.1.1
flask-cors==5.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.2.1
pycparser==2.22
tinycss2==1.4.0
webencodings==0.5.1
Werkzeug==3.1.3
